Update readme "Configuration" section to make the VS Code settings more explicit
I was very pleased to discover this extension, but struggled for a while to figure out the configuration. I didn't read carefully enough, and got confused between mypy.ini settings and VS Code settings.

This patch attempts to make the readme a little more clear about which settings are for VS Code. Additionally, I listed all of the extension config settings in the readme. If that seems undesirable, I suggest at the very least being clear that `mypy.targets` is a VS Code setting, and not a mypy setting, which is where I got tripped up.

Thanks for the great project!